+++ src/settings/ActionProfiles/ActionProfilesForm.js
@@ -94,8 +94,17 @@ export const ActionProfilesFormComponent = ({
           ACTION_TYPES_SELECT.UPDATE.type,
         ]);
       }
-      case ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES.MARC_AUTHORITY.type: {
-        return pick(ACTION_TYPES_SELECT, ACTION_TYPES_SELECT.UPDATE.type);
+      case ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES.INSTANCE.type: {
+        return pick(ACTION_TYPES_SELECT, [
+          ACTION_TYPES_SELECT.CREATE.type,
+          ACTION_TYPES_SELECT.UPDATE.type,
+        ]);
+      }
+      case ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES.HOLDINGS.type: {
+        return pick(ACTION_TYPES_SELECT, [
+          ACTION_TYPES_SELECT.CREATE.type,
+          ACTION_TYPES_SELECT.UPDATE.type,
+        ]);
       }
       default: {
         return ACTION_TYPES_SELECT;
@@ -126,7 +135,10 @@ export const ActionProfilesFormComponent = ({
       }
       case ACTION_TYPES_SELECT.CREATE.type:
       default: {
-        return ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES;
+        return omit(ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES, [
+          ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES.MARC_HOLDINGS.type,
+          ACTION_PROFILES_FORM_FOLIO_RECORD_TYPES.ORDER.type,
+        ]);
       }
     }
   };